rename the comfyui files and new readme
